% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello-world.R
\name{helloWorld}
\alias{helloWorld}
\title{Hello world application}
\usage{
helloWorld()
}
\value{
Nothing
}
\description{
Simplest shiny application available in shiny galary
}
\details{
Used for demo R package with shiny tools
}
\examples{
if (interactive()) {
  helloWorld()
}

}
\keyword{Hello}
\keyword{app}
\keyword{world}
