% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means-clustering.R
\name{kmeansClustering}
\alias{kmeansClustering}
\title{Iris k-means clustering}
\usage{
kmeansClustering()
}
\value{
Nothing
}
\description{
Simplest shiny application available in shiny galary
}
\details{
Used for demo R package with shiny tools
}
\examples{
if (interactive()) {
  kmeansClustering()
}

}
